requirements.txt
Flask==0.10.1
Jinja2==2.7.3
MarkupSafe==0.23
Werkzeug==0.9.6
elasticsearch==1.2.0
itsdangerous==0.24
pyasn1==0.1.7
pysmb==1.1.13
python-nmap==0.3.4
pywin32==219
urllib3==1.9.1
wsgiref==0.1.2
